Changed dec to int

+++ Reciply.API/Models/Ingredient.cs
@@ -5,7 +5,7 @@
         public int IngredientId { get; set; }
 
         public string Name { get; set; }
-        public decimal Quantity { get; set; }
+        public int Quantity { get; set; }
         public string Unit { get; set; }
 
         public int RecipeId { get; set; }
